before logging in, the auth store now resets itself, same for on init
